大改版前的保存

+++ src/ndsl/net/Multiplexer.cc
@@ -129,7 +129,7 @@ void Multiplexer::dispatch(void *p)
     Multiplexer *pthis = static_cast<Multiplexer *>(p);
 
     //有不完整头部出现时，将其复制到msghead开始处，然后调用onrecv从残缺头部开始放
-    if (pthis->rlen_ < sizeof(struct Message) && pthis->left_ == 0) {
+    if ((size_t) pthis->rlen_ < sizeof(struct Message) && pthis->left_ == 0) {
         if (pthis->location_ != pthis->msg_)
             memcpy(pthis->msg_, pthis->location_, pthis->rlen_);
         pthis->location_ = pthis->msg_ + pthis->rlen_;
